+++ refena/test/unit/redux/redux_notifier_lifecycle_test.dart
@@ -142,7 +142,6 @@ void main() {
       ProviderDisposeEvent(
         debugOrigin: container,
         provider: provider,
-        notifier: notifier,
       ),
     ]);
   });